+++ server/core_notification.go
@@ -101,7 +101,7 @@ func NotificationList(ctx context.Context, logger *zap.Logger, db *sql.DB, userI
 SELECT id, subject, content, code, sender_id, create_time
 FROM notification
 WHERE user_id = $1`+cursorQuery+`
-ORDER BY create_time ASC`+limitQuery, params...)
+ORDER BY create_time ASC, id ASC`+limitQuery, params...)
 
 	if err != nil {
 		logger.Error("Could not retrieve notifications.", zap.Error(err))